requirements.txt
aiohttp==3.7.4.post0
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
branca==0.3.1
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
dj-database-url==0.5.0
Django==3.2.6
django-crispy-forms==1.12.0
django-heroku==0.3.1
django-ipware==4.0.0
DriveDownloader==1.3.0
folium==0.12.1
geographiclib==1.52
geoip2==4.2.0
geopy==2.2.0
googledrivedownloader==0.4
gunicorn==20.1.0
idna==3.2
Jinja2==3.0.1
MarkupSafe==2.0.1
maxminddb==2.0.3
multidict==5.1.0
numpy==1.21.2
psycopg2==2.9.1
PySocks==1.7.1
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
tqdm==4.62.2
typing-extensions==3.10.0.0
urllib3==1.26.6
whitenoise==5.3.0
yarl==1.6.3
